Restruccturacion de main
