--- src/components/Main.css
@@ -1,55 +1,0 @@
-.note-area-container {
-    border: 1px solid black;
-    width: 80%;
-    align-items: center;
-    border-radius: 0.5em;
-    margin-top: 1em;
-    padding: 0.5em;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.note-area-title {
-    border: none;
-    font-size: 1.3em;
-    margin-bottom: .75em;
-}
-
-.note-area-title:focus-visible {
-    outline: none;
-}
-
-.note-area-textarea {
-    width: 90%;
-    border: none;
-    height: 6em;
-    font-size: 1em;
-}
-
-.note-area-textarea:focus-visible {
-    outline: none;
-}
-
-.close-note-area {
-    text-align: right;
-}
-
-.close-note-area button {
-    border: none;
-    background: none;
-    color: #0c0b0b;
-    font-size: 1em;
-    padding: 0.5em 0.75em;
-}
-
-.minimized .note-area-title {
-    display: none;
-}
-
-.minimized .note-area-textarea {
-    height: 1em;
-}
-
-.minimized .close-note-area {
-    display: none;
-}
