+++ src/components/Notes.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import './Notes.css';
 import { MyUser } from "../models/MyUser";
 import Header from "./Header";
-import Main from "./Main";
+import CreateNotePanel from "./CreateNotePanel";
 
 type NotesProps = {
   user: MyUser
@@ -12,7 +12,7 @@ export default function Notes({ user }: NotesProps) {
   return (
     <React.Fragment>
       <Header user={user}></Header>
-      <Main user={user}></Main>
+      <CreateNotePanel user={user}></CreateNotePanel>
 
     </React.Fragment>
   )
